CMake: Added source groups for agent.
This commit will continue our effort to break libmesos into a modular
set of libraries that can be built and linked independently, and away
from the approach of creating a single, monolithic libmesos library.
This effort is tracked in MESOS-3542.

This commit will specifically address the problem of breaking agent code
into a self-contained source group, which can then be exported as its
own library. This code also has the benefit of making IDEs display the
agent code in a hierarchical directory structure.

Review: https://reviews.apache.org/r/55601