Windows: Added parent hooks to subprocess.
The subprocess API exposes various hooks to customize behavior as we
create the child process. Currently, these hooks are unimplemented in
the Windows codepaths. This commit will implement the parent hooks --
which are executed after the child process is created, but before it
begins execution -- on the Windows codepath.

Review: https://reviews.apache.org/r/55312