CMake: Added `test` target.
This commit will introduce a `test` target to the CMake build system.
The semantics of this target are very similar to the autotools build
system, in that they build the tests, but do not run them.

Accomplishing this is somewhat complex in CMake, because CMake expects
to control the `test` target itself. We work around this by (1)
silencing the warning that `test` is a reserved target, (2) removing the
call to `enable_testing` that sets up CTest, and (3) adding our own
`test` target. As an additional insurance policy, we error out if any of
the `BUILD_TESTING` flags are defined, which would indicate that
`include(CTest)` was called.

Review: https://reviews.apache.org/r/55637