Copy `PosixLauncher` to create the `WindowsLauncher`.
Instead of deriving `WindowsLauncher` from `PosixLauncher`,
this commit implements a separate `WindowsLauncher` derived
from `Launcher` parallel to the `PosixLauncher`.

The purpose of this is to setup for refactoring the `WindowsLauncher`
into a `Launcher` that uses Windows' "Job Objects,"
which are similar to Linux's cgroups, and enable use to reason
about a group of processes. It is necessary to do this in the
`WindowsLaucher` because the current implementation uses the POSIX
idea of a process hierarchy to list and kill all processes
associated with a task. We have found this model to not work on
Windows, and the solution is to use Job Objects.

Because this is preparation work, it is a straight copy of the
current "working" implementation. That is, the `PosixLauncher`
has been copied to create the `WindowsLauncher`, instead of
the latter deriving the former.