WIP: Fix several scoping issues for Windows PCH.
This commit needs to be completely redone. It's a work-in-progress
to enable pre-compiled headers via the Cotire module on Windows.
The inclusion of the PCH uncovered multiple namespace collision
issues. Some were solved by excluding certain headers from the PCH,
others were solved by qualifying the namespaces in this commit.
However, these are actual bugs that need to be tracked,
fixed correctly, and the commit cleaned up (currently made with
--no-verify because of line length issues).