WIP: Fix several scoping issues for Linux PCH.
This starts to fix several namespaces issues uncovered by the inclusion
of PCH for Linux, similarly to the previous commit. It is purposefully
kept separate as this was uncovered on Linux, and the former on Windows.
These particular changes can probably be used as-is as it was just
global qualification (prepend `::`) plus a style fix from C style cast
to `static_cast`.