WIP: "Fix" stout/set.hpp.
This is another issue uncovered by the attempt to add PCH to the
Linux build. This is *not* a proper a fix. The code as-is is a
standards violation as it adds functions to the `std` namespace,
which is (a) not allowed and (b) breaks ADL, causing the `operator+()`
to not be found when including certain headers. An issue for this
should be filed as this must be fixed another way.