CMake: Enable `/debug:fastlink` on Windows.
A previous commit attempted to enable this option, but it was not
actually being used by the linker. This patch correctly adds the option
to each linker flag variable for shared libraries, modules, and
executables, and only for the debug configuration. This results in much
faster link times.

Also, the `/zc:inline` option was removed, as it is only relevant in
optimized builds (i.e. release configurations).