Windows: Explicitly set `cwd` for `CreateProcess`.
This change uses `os::getcwd()` to obtain the current working directory
without the long path prefix `\\?\`. This does not change the current
behavior otherwise, as `nullptr` defaulted to the process's `cwd`.

Also, this patch removes the unsafe and implicit C-style casts. Where
the `processEnvironment` was not a `const void*`, an actual non-const
data-structure is supplied instead of `std::wstring::data()`.