Removed `ENABLE_OPTIMIZE` CMake option.
This option was added as a mapping from the Autotools
`--enable-optimize` flag. However, it does not work as expected, since
CMake automatically adds platform-specific optimization flags depending
on the build configuration. That is, Debug builds are not optimized, and
Release builds are optimized. Therefore, all this option did was to
unnecessarily add optimizations to the Debug build, and fail to disable
optimizations for the Release build.