Windows: Fixed OpenSSL linking to prefer static libraries.
Note that this is not done in 3rdparty for OpenSSL, as on Linux, shared
libraries are generally preferred, and so this option can be set
on-the-fly (and is correctly passed to dependencies). However, on
Windows OpenSSL has to be installed separately. So it is simpler for
deployment to continue sticking with static libraries for our third
party dependencies.