Windows: Fixed `fs::list` to return full paths.
This resolves MESOS-7803, where `fs::list` only returned the name of the
found files, dropping the initial path components. That is, for a
directory `foo` with the file `bar` in it, and the pattern
`fs::list("foo/*")`, it was incorrectly returning just `bar`, not
`foo/bar`. This was fixed by getting the `dirname` of the pattern, and
rejoining it with each found file name before it is returned.