Windows: Implemented `os::cloexec` family for file descriptors.
The intent of `os::cloexec` (that is, the intent of the `CLOEXEC` flags
on Linux) is to disable the inheritance of handles. On Linux, the flag
specifically means to not duplicate the file descriptor on calls to the
`exec` family of functions (which spawn new processes). On Windows, the
logical equivalent is to set `HANDLE_FLAG_INHERIT` to `0`, which means
to not allow the handles (file descriptors) to be inherited (duplicated)
in child processes (exec).

Note that it is not possible to change the inheritance behavior of an
already-existing socket on Windows, so we log a warning and return
nothing. However, inheritance can be controlled at socket creation, and
so `os::isCloexec` supports sockets too.