Windows: Disabled `O_CLOEXEC` semantic mapping on Windows.
Previously we mapped `O_CLOEXEC` to `O_NOINHERIT`, which is close to the
intended semantics, but not close enough. Instead, we want all Windows
handles by default to be non-inheritable, and not tie any Windows
inheritance semantics to the close-on-exec semantics. So, we define
`O_CLOEXEC` on Windows to be `0`, and remove the logging from the
`os::cloexec` family of functions because do not ever intend to
implement them.

We achieve the "close" part of the semantics by making all handles
non-inheritable, as the intent is to avoid leaking handles.

Review: https://reviews.apache.org/r/65403