Windows: Fixed `os::dup()`.
For file descriptors, this was using the CRT API `::_dup()`. However, it
is more correct to use the Windows API `::DuplicateHandle` as it gives
us full control over inheritance semantics and access level.

For sockets, we were still using a non-Unicode API, and moreover, had
disabled the warning. Now we use the Unicode API.