Windows: Abstracted CRT, Handle, and Socket types in WindowsFD.
This patch refactors our implementation of `WindowsFD`. The original
design held a union of the primitive types represented (`int`, `HANDLE`,
`SOCKET`), and implemented each method via a switch statement. This
caused the logic to be mixed. The new design now has three separate
types, `IntFD`, `HandleFD`, and `SocketFD`, each of which is
self-contained and describes how it should be constructed and casted. We
then use a composition pattern to create the `WindowsFD` abstraction
known as `int_fd` elsewhere. Each method of `WindowsFD` simply
dispatches to the underlying type.

Part of this new design was separating `CrtFD` from `HandleFD`. While
their interfaces are similar, they are constructed differently.  For
this reason, they are separated into two types. When we start replacing
`int_fd` with RAII types, this design will give us a clean separation of
a Windows file `HANDLE` and a Windows CRT file `int`, the latter of
which we plan to eventually remove.

Previously, we had a subtle bug where a `WindowsFD` could be constructed
from a seemingly valid `HANDLE`, but the constructor silently failed to
open the corresponding CRT `int` file descriptor. We now `CHECK` the
validity of the objects on construction.

The comparison operators for `WindowsFD` have been simplified. Instead
of providing an explicit `int crt()` function on the `IntFD` and
`HandleFD` types, we instead define `operator int()` for each of the
three types. The important part of the definition is that it returns
`-1` if the corresponding primitive is invalid (e.g.
`INVALID_HANDLE_VALUE` or `nullptr` for `HandleFD` and `INVALID_SOCKET`
for `SocketFD`). We are then able to overload the equality and
comparison operators of `WindowsFD` to always compare as an `int`. With
this design, a `WindowsFD` can be compared to `== -1` or be `>= 0` etc.
as is POSIX fashion, and be reasonably checked for validity. Similarly,
a `CrtFD` can be compared to a `HandleFD` and they are checked by their
underlying primitive `int crt`, providing semantically equivalent
behavior as if the abstraction didn't exist.