Windows: Fixed `os::ftruncate()` to use `SetEndOfFile()`.
This previously used the CRT API `_chsize_s()`, which required a CRT
integer file descriptor. Instead, we can achieve the same behavior by
first using `os::lseek()` (which uses `SetFilePointerEx()`) to seek
`length`, and then use `SetEndOfFile()` to truncate. The only
difference is that the file is not filled with null bytes when
expanded, but we do not seem to rely on this behavior.

Review: https://reviews.apache.org/r/66455