Ported (some) Python 2 support scripts to Python 3.
According to https://docs.python.org/3/howto/pyporting.html, we can
make our support scripts compatible with both Python 2 and 3 provided
we only worry about supporting Python 2.7.

Using the Python-Future quick start guide
http://python-future.org/quickstart.html, I installed a Python 3.6.5
environment, ran `pip install future`, and then ran `futurize -w` on
the modified scripts.

In my unchanged system environment, attempting to use the scripts
resulted in a traceback with the error `ImportError: No module named
future`, since the `future` package is sadly not installed by default.
On Ubuntu 17.10, I was able to `apt install python-future` to add it
to my system's Python 2 installation, so would be a new requirement
despite the Python 2 support.

The ported scripts were chosen in order to support using Python 3 on
the Windows build bot, so that author names with non-ASCII characters
can be used. The rest of the scripts should follow, but are not
necessary, and require more extensive testing and fixes.

A manual change was made to `post-reviews` to deal with the stdout of
child processes. In `execute` we explicitly decode the stdout to a
string to match the existing expectation (in Python 2, it was already
a string, but in Python 3 bytes and strings are different types). When
printing the stdout of a child process, we use `sys.stdout.buffer` to
print bytes instead of a string in order to preserve ANSI color
escapes in the output. This also fixes an existing bug on Windows
where the script would not correctly display colors.

Review: https://reviews.apache.org/r/66964