CMake: Removed explicit linkage of libapr.
Because SVN now includes APR as an interface dependency, the explicit
linkage to APR in the stout interface is not necessary. Instead, APR
is included transitively when SVN is included.

This fixes the unexpected behavior on Windows were APR is built and
linked against despite not being used at all since SVN isn't used.

Review: https://reviews.apache.org/r/67535/