Windows: Fixed a bug when `executorId` contains a `:`.
Per MESOS-9109, some frameworks, such as Chronos, use a `:` in their
`executorId`, but this is a reserved character on Windows, and when
its used as part of the path on disk results in an OS error.
Unfortunately, we need to remain backward-compatible with existing
frameworks, so now we escape `:` as `%3A` when writing to disk, and
unescape it when parsing.

Review: https://reviews.apache.org/r/68297/